requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
bcrypt==5.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.1
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.128.0
h11==0.16.0
idna==3.11
openpyxl==3.1.5
passlib==1.7.4
pillow==12.1.0
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
razorpay==2.0.0
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
gunicorn
